clothing_scraper/requirements.txt
pyppeteer
psycopg2-binary
fastapi
uvicorn[standard]
undetected-chromedriver
python-2captcha
scrapy